Refactor token handling in ask_your_token.py and update bot.py to improve command handling and logging

+++ keys/tools/ask_your_token.py
@@ -2,7 +2,7 @@ import os
 
 
 def ask_your_token():
-    token= input("Enter your telegram bot token: ")
-    with open('./keys/_token.py','w') as file:
-        file.write(f"_token= '{token}'\n") 
-# ask_your_token()
+    if not os.path.exists('./keys/_token.py'):
+        token= input("Enter your telegram bot token: ")
+        with open('./keys/_token.py','w') as file:
+            file.write(f"_token= '{token}'\n")
